only delete when there is set value for timestamp
git-svn-id: svn+ssh://svn.wikia-code.com/svn/wikia/trunk@29179 e846bd91-c042-42a7-a218-4b65e2c1904d

+++ includes/wikia/RiakCache.php
@@ -136,7 +136,7 @@ class RiakCache extends BagOStuff {
 		if( is_array( $data ) ) {
 			$value     = array_shift( $data );
 			$timestamp = array_shift( $data );
-			if( $timestamp < time() ) {
+			if( $timestamp && $timestamp < time() ) {
 				$this->delete( $key );
 				$value = false;
 			}
